Refactor where doc is obtained
